refactored code to eliminate folders and files.
