+++ requirements.txt
@@ -1,64 +1,54 @@
 annotated-types==0.7.0
-anthropic==0.30.1
+anthropic==0.31.0
 anyio==4.4.0
-asttokens==2.4.1
 attrs==23.2.0
 backoff==2.2.1
 certifi==2024.7.4
 charset-normalizer==3.3.2
-comm==0.2.2
-debugpy==1.8.2
-decorator==5.1.1
+click==8.1.7
 distro==1.9.0
-exceptiongroup==1.2.1
-executing==2.0.1
+dnspython==2.6.1
+email_validator==2.2.0
+fastapi==0.111.0
+fastapi-cli==0.0.4
 filelock==3.15.4
 fsspec==2024.6.1
 h11==0.14.0
 httpcore==1.0.5
+httptools==0.6.1
 httpx==0.27.0
 huggingface-hub==0.23.4
 idna==3.7
-ipykernel==6.29.5
-ipython==8.26.0
-jedi==0.19.1
 Jinja2==3.1.4
 jiter==0.5.0
 jsonlines==4.0.0
-jupyter_client==8.6.2
-jupyter_core==5.7.2
-langfuse==2.38.0
-lightrag==0.0.0a8
+lightrag==0.0.0a16
+markdown-it-py==3.0.0
 MarkupSafe==2.1.5
-matplotlib-inline==0.1.7
-nest-asyncio==1.6.0
+mdurl==0.1.2
 numpy==1.26.4
-packaging==23.2
-parso==0.8.4
-pexpect==4.9.0
-platformdirs==4.2.2
-prompt_toolkit==3.0.47
-psutil==6.0.0
-ptyprocess==0.7.0
-pure-eval==0.2.2
+orjson==3.10.6
+packaging==24.1
 pydantic==2.8.2
 pydantic_core==2.20.1
 Pygments==2.18.0
-python-dateutil==2.9.0.post0
 python-dotenv==1.0.1
+python-multipart==0.0.9
 PyYAML==6.0.1
-pyzmq==26.0.3
 regex==2024.5.15
 requests==2.32.3
-six==1.16.0
+rich==13.7.1
+shellingham==1.5.4
 sniffio==1.3.1
-stack-data==0.6.3
+starlette==0.37.2
 tiktoken==0.7.0
 tokenizers==0.19.1
-tornado==6.4.1
 tqdm==4.66.4
-traitlets==5.14.3
+typer==0.12.3
 typing_extensions==4.12.2
+ujson==5.10.0
 urllib3==2.2.2
-wcwidth==0.2.13
-wrapt==1.16.0
+uvicorn==0.30.1
+uvloop==0.19.0
+watchfiles==0.22.0
+websockets==12.0
